Extract event bus persistence into integrations to fix layer violation
This commit resolves a sentrux/import-linter architecture violation where the
core `event_bus.handlers` directly imported `app.crud.chat_message` and
`app.crud.workspace`. To restore proper layering (`be-core -> be-crud`), the
persistence and turn execution functions were extracted to a new dependency
module `app/integrations/event_bus_deps.py` and are now injected into
`AgentHandler` upon initialization in `main.py`. The legacy grandfathered
exceptions in `.importlinter` have been removed.

- Add `RunAgentTurnFn` and `PersistAssistantResponseFn` protocols.
- Inject these callables into `AgentHandler`.
- Move `_run_agent_turn` and `_persist_assistant_response` out of core.
- Update `backend/main.py` and tests.
- Remove `.importlinter` exception.
