+++ backend/app/core/event_bus/handlers.py
@@ -22,8 +22,7 @@ from __future__ import annotations
 
 import logging
 import uuid
-from collections.abc import Iterable
-from pathlib import Path
+from collections.abc import Awaitable, Callable, Iterable
 from typing import Any
 
 from app.core.event_bus.bus import Event
@@ -32,10 +31,13 @@ from app.core.event_bus.types import (
     ScheduledEvent,
     WebhookEvent,
 )
-from app.core.providers import default_model, resolve_llm
-from app.db import async_session_maker
 
 logger = logging.getLogger(__name__)
+
+# Injected dependencies to avoid core -> crud layering violations
+RunAgentTurnFn = Callable[[str, uuid.UUID], Awaitable[str]]
+PersistAssistantResponseFn = Callable[[uuid.UUID, uuid.UUID, str, str], Awaitable[None]]
+
 
 # How many characters of the webhook payload to flatten into the prompt.
 # Past this we truncate so a noisy GitHub payload can't blow the model
@@ -68,8 +70,16 @@ class AgentHandler:
     handler that persists a conversation row per fire is a follow-on.
     """
 
-    def __init__(self, *, default_user_id: uuid.UUID | None = None) -> None:
+    def __init__(
+        self,
+        *,
+        default_user_id: uuid.UUID | None = None,
+        run_turn_fn: RunAgentTurnFn | None = None,
+        persist_response_fn: PersistAssistantResponseFn | None = None,
+    ) -> None:
         self._default_user_id = default_user_id
+        self._run_turn_fn = run_turn_fn
+        self._persist_response_fn = persist_response_fn
 
     def register(self, bus: Any) -> None:
         """Attach the agent handler to the bus.
@@ -142,8 +152,12 @@ class AgentHandler:
                 originating_event_id,
             )
             return
+        if self._run_turn_fn is None:
+            logger.error("AGENT_HANDLER_NO_RUN_TURN_FN originating_event_id=%s", originating_event_id)
+            return
+
         try:
-            text = await _run_agent_turn(prompt=prompt, user_id=user_id)
+            text = await self._run_turn_fn(prompt, user_id)
         except Exception:
             logger.exception(
                 "AGENT_HANDLER_RUN_FAILED originating_event_id=%s user_id=%s",
@@ -159,12 +173,18 @@ class AgentHandler:
             )
             return
         if target_conversation_id is not None:
-            await _persist_assistant_response(
-                conversation_id=target_conversation_id,
-                user_id=user_id,
-                text=text,
-                originating_event_id=originating_event_id,
-            )
+            if self._persist_response_fn is not None:
+                await self._persist_response_fn(
+                    target_conversation_id,
+                    user_id,
+                    text,
+                    originating_event_id,
+                )
+            else:
+                logger.warning(
+                    "AGENT_HANDLER_NO_PERSIST_FN originating_event_id=%s",
+                    originating_event_id,
+                )
         # Lazy import — keeps the handler module decoupled from the
         # bus-publish helper to avoid a circular import surface.
         from app.core.event_bus.global_bus import (  # noqa: PLC0415
@@ -323,141 +343,3 @@ def _flatten_dict(
             lines.append(f"{full}: {_format_leaf(value)}")
 
 
-async def _run_agent_turn(*, prompt: str, user_id: uuid.UUID) -> str:
-    """Stream one provider turn and return the concatenated assistant text.
-
-    Resolves the catalog default model + the user's workspace + the
-    standard tool composition.  Skips workspace tools when the user
-    hasn't completed onboarding (no default workspace) — webhook /
-    scheduled traffic shouldn't be gated on the onboarding flow.
-    """
-    # All imports are lazy so the bus module can be loaded without
-    # pulling in the chat router's heavy dependency tree.
-    from app.core.agent_loop.tools import build_agent_tools  # noqa: PLC0415
-    from app.core.governance.permissions import (  # noqa: PLC0415
-        PermissionContext,
-        build_default_permission_check,
-    )
-    from app.core.governance.workspace_context import (  # noqa: PLC0415
-        load_workspace_context,
-    )
-    from app.crud.workspace import get_default_workspace  # noqa: PLC0415
-
-    async with async_session_maker() as session:
-        workspace = await get_default_workspace(user_id, session)
-
-    workspace_root: Path | None = Path(workspace.path) if workspace is not None else None
-    workspace_ctx = load_workspace_context(workspace_root) if workspace_root is not None else None
-    system_prompt = workspace_ctx.system_prompt if workspace_ctx is not None else None
-    enabled_tools = workspace_ctx.enabled_tools if workspace_ctx is not None else None
-
-    agent_tools = (
-        build_agent_tools(
-            workspace_root=workspace_root,
-            user_id=user_id,
-            send_fn=None,
-            surface="webhook",
-        )
-        if workspace_root is not None
-        else []
-    )
-
-    from app.core.agent_loop.types import (  # noqa: PLC0415
-        PermissionCheckFn,
-        PermissionCheckResult,
-    )
-
-    permission_check_fn: PermissionCheckFn | None = None
-    if workspace_root is not None:
-        permission_context = PermissionContext(
-            user_id=str(user_id),
-            workspace_root=workspace_root,
-            conversation_id=str(uuid.uuid4()),
-            surface="webhook",
-            enabled_tools=enabled_tools,
-        )
-        gate = build_default_permission_check()
-
-        async def permission_check_for_handler(
-            tool_name: str, arguments: dict[str, Any]
-        ) -> PermissionCheckResult:
-            decision = await gate(tool_name, arguments, permission_context)
-            return PermissionCheckResult(
-                allow=decision.allow,
-                reason=decision.reason,
-                violation_type=decision.violation_type,
-            )
-
-        permission_check_fn = permission_check_for_handler
-
-    # resolve_llm does not accept user_id; workspace_root carries the
-    # per-user key resolution upstream. Kept for call-site symmetry.
-    _ = user_id
-    provider = resolve_llm(
-        default_model().id,
-        workspace_root=workspace_root,
-    )
-
-    accumulated: list[str] = [
-        stream_event.get("content", "")
-        async for stream_event in provider.stream(
-            prompt,
-            uuid.uuid4(),
-            user_id,
-            history=[],
-            tools=agent_tools or None,
-            system_prompt=system_prompt,
-            permission_check=permission_check_fn,
-        )
-        if stream_event.get("type") == "delta"
-    ]
-    return "".join(accumulated).strip()
-
-
-async def _persist_assistant_response(
-    *,
-    conversation_id: uuid.UUID,
-    user_id: uuid.UUID,
-    text: str,
-    originating_event_id: str,
-) -> None:
-    """Write the agent's response into a chat conversation as a finalised turn.
-
-    Lazy imports keep ``event_bus.handlers`` from pulling the
-    chat-message CRUD into its module-load graph — the sentrux layer
-    rule wants core code to avoid hard imports of ``app.crud.*``.
-
-    Failures are logged and swallowed: a persistence error must not
-    break the Telegram fan-out that follows in the caller. The row is
-    written via the same helpers the web chat router uses, so the
-    UI's existing ``GET .../messages`` path picks it up immediately.
-    """
-    from app.crud.chat_message import (  # noqa: PLC0415
-        append_assistant_placeholder,
-        finalize_assistant_message,
-    )
-
-    try:
-        async with async_session_maker() as session:
-            placeholder = await append_assistant_placeholder(
-                session,
-                conversation_id=conversation_id,
-                user_id=user_id,
-            )
-            await finalize_assistant_message(
-                session,
-                message_id=placeholder.id,
-                content=text,
-                thinking=None,
-                tool_calls=None,
-                timeline=None,
-                thinking_duration_seconds=None,
-                assistant_status="complete",
-            )
-            await session.commit()
-    except Exception:
-        logger.exception(
-            "AGENT_HANDLER_PERSIST_FAILED conversation_id=%s originating_event_id=%s",
-            conversation_id,
-            originating_event_id,
-        )
